Fix hadolint reported problem and incorrect build variable for iwyu
Changes:
Reworks the iwyu installation to avoid cd commands - to fix a hadolint warning.
IWYU_LLVM_ROOT_PATH was being set but was unused. CMAKE_PREFIX_PATH should be used instead (with the same value) for IWYU >=0.11

Tests:
Checked that the container can still build and run, that the unused variable warning is gone and that hadolint gives a clean report.